Updated tests and a couple other things
A few things: Do we need index for GamesController?
Do we ever use friends method in the Users Controller? It seems to run fine without it.
Buttons on home page need to be updated to bring user to the right path, not just user/2 but user/friends/2 or similar - for clicking on a friend.